peepee poopoo

+++ src/utils.sh
@@ -9,9 +9,24 @@ function mkdirprint(){
 }
 
 function cpprint(){
-    if [[ $CHECKING -eq 0 ]]; then
-        cp -a "$1" "$2";
+
+    FILE_MODE_DATE=$(stat -c %Y "$1")
+    if [[ -f $2 ]]; then
+        BAK_FILE_DATE=$(stat -c %Y "$2")
+        if [[ "$FILE_MODE_DATE" -gt "$BAK_FILE_DATE" ]]; then
+            if [[ $CHECKING -eq 0 ]]; then
+                cp -a "$1" "$2";
+            fi
+            echo "cp -a $1 $2"
+        else
+            echo "$2 is newer than $1"
+        fi
+    else
+        if [[ $CHECKING -eq 0 ]]; then
+            cp -a "$1" "$2";
+        fi
+        echo "cp -a $1 $2"
+        return 0;
     fi
-    echo "cp -a $1 $2"
     return 0;
 }
